component re-do and nav bar cleanup
