+++ app/components/Manager.js
@@ -4,15 +4,16 @@ var helpers = require("./utils/helpers");
 var Manager = React.createClass({
 
     getInitialState: function() {
-        return { 
-            username: "" 
+        return {
+            username: "",
+            picture: ""
         };
     },
 
     componentDidMount: function() {
        helpers.getCurrentUser().then(function(response) {
           if (response !== this.state.username) {
-            this.setState({ username: response.data.username });
+            this.setState({ picture: response.data.picture, username: response.data.username });
           }
         }.bind(this));
     },
@@ -20,16 +21,17 @@ var Manager = React.createClass({
     render: function() {
         return (
             <div>
+                <ul id="dropdown1" className="dropdown-content">
+                    <li><a className="black-text" href="/logout">Logout<i className="material-icons right">exit_to_app</i></a></li>
+                </ul>
                 <nav>
                     <div className="nav-wrapper grey lighten-5">
-                        <a href="/manager" className="brand-logo blue-text text-darken-1"><img id="logo" src="/assets/images/logo.png"/><span className="hide-on-med-and-down">Schedulr [{this.state.username}]</span></a>
+                        <a href="/manager" className="brand-logo blue-text text-darken-1"><img id="logo" src="/assets/images/logo.png"/><span className="hide-on-med-and-down">Schedulr</span></a>
                         <a href="/" data-activates="slide-out" className="button-collapse blue-text text-darken-1"><i className="material-icons">menu</i></a>
                         <ul className="right hide-on-med-and-down">
                             <li><a className="black-text" href="/manager/employeeAll">Employee Management<i className="material-icons right">group</i></a></li>
                             <li><a className="black-text" href="/manager/schedulesCreate">Schedules<i className="material-icons right">access_time</i></a></li>
-                            <li><a className="black-text" href="/logout">Logout<i className="material-icons right">exit_to_app</i></a></li>
-                           
-                           
+                            <li><a className="dropdown-button black-text" href="#" data-activates="dropdown1" data-beloworigin="true" data-hover="true">{this.state.username}<img id="picture" className="circle circle-small" src={this.state.picture}/></a></li>
                         </ul>
                         <ul id="slide-out" className="side-nav">
                             <li>
@@ -37,7 +39,7 @@ var Manager = React.createClass({
                                     <div className="background">
                                         <img src="http://materializecss.com/images/office.jpg"/>
                                     </div>
-                                    <a><img className="circle" src="/assets/images/logo.png"/></a>
+                                    <a><img className="circle" src={this.state.picture}/></a>
                                     <a><span className="white-text">Company Name</span></a>
                                     <a><span className="white-text name">{this.state.username}</span></a>
                                 </div>
@@ -50,8 +52,6 @@ var Manager = React.createClass({
                 </nav>
                 <div className="container">
                     {this.props.children}
-                    
-                            
                 </div>
             </div>
         );
