+++ app/components/children/AnnouncementsBuild.js
@@ -1,0 +1,13 @@
+var React = require("react");
+
+var AnnouncementsView = React.createClass({
+    render: function() {
+        return (
+            <div>
+                <p>Build announcements here</p>
+            </div>
+        );
+    }
+});
+
+module.exports = AnnouncementsView;
