+++ app/components/children/AnnouncementsView.js
@@ -1,0 +1,13 @@
+var React = require("react");
+
+var AnnouncementsView = React.createClass({
+    render: function() {
+        return (
+            <div>
+                <p>View announcements here</p>
+            </div>
+        );
+    }
+});
+
+module.exports = AnnouncementsView;
